+++ test/e2e/okta/okta-oauth-google.spec.ts
@@ -12,9 +12,8 @@ import {JWTPayload, OktaOAuthProvider, OktaUser} from 'services/okta.interfaces'
 import {
     getMockOktaUser,
     mockOktaCreateUser,
-    mockOktaListUsers,
+    mockOktaGetUserByEmail,
     mockOktaSendActivationEmail,
-    mockOktaUpdateUser
 } from './okta.mocks';
 
 chai.should();
@@ -97,12 +96,7 @@ describe('[OKTA] Google auth endpoint tests', () => {
             providerId,
         });
 
-        mockOktaListUsers({
-            limit: 1,
-            search: `(profile.provider eq "${OktaOAuthProvider.GOOGLE}") and (profile.providerId eq "${providerId}")`
-        }, [user]);
-
-        mockOktaUpdateUser(user, { email: 'john.doe@vizzuality.com' });
+        mockOktaGetUserByEmail(user.profile);
 
         nock('https://www.googleapis.com')
             .get('/oauth2/v1/userinfo')
@@ -144,10 +138,16 @@ describe('[OKTA] Google auth endpoint tests', () => {
             providerId,
         });
 
-        mockOktaListUsers({
-            limit: 1,
-            search: `(profile.provider eq "${OktaOAuthProvider.GOOGLE}") and (profile.providerId eq "${providerId}")`
-        }, []);
+        // Mock user not found
+        nock(config.get('okta.url'))
+            .get(`/api/v1/users/${user.profile.email}`)
+            .reply(404, {
+                'errorCode': 'E0000007',
+                'errorSummary': `Not found: Resource not found: ${user.profile.email} (User)`,
+                'errorLink': 'E0000007',
+                'errorId': 'oaeM-EhNh-aRXmmjoxRYFUgLQ',
+                'errorCauses': []
+            });
 
         mockOktaCreateUser(user, {
             email: 'john.doe@vizzuality.com',